Expand save dialog and preferred save location option in the setting

+++ app/src/main/java/de/ur/mi/audidroid/viewmodels/RecordViewModel.kt
@@ -14,6 +14,7 @@ import androidx.lifecycle.MutableLiveData
 import com.google.android.material.snackbar.Snackbar
 import de.ur.mi.audidroid.R
 import de.ur.mi.audidroid.models.EntryEntity
+import de.ur.mi.audidroid.models.LabelAssignmentEntity
 import de.ur.mi.audidroid.models.MarkerTimeRelation
 import de.ur.mi.audidroid.models.Repository
 import java.io.File
@@ -119,7 +120,7 @@ class RecordViewModel(private val dataSource: Repository, application: Applicati
     }
 
     fun cancelRecord() {
-        if (recorderInitialized) {
+        if (recorderInitialized && createDialog.value == false) {
             showSnackBar(R.string.record_removed)
             File(tempFile).delete()
             endRecordSession()
@@ -173,19 +174,26 @@ class RecordViewModel(private val dataSource: Repository, application: Applicati
         mediaRecorder.resume()
     }
 
-    fun getNewFileFromUserInput(nameInput: String?, pathInput: String?) {
+    fun getNewFileFromUserInput(
+        nameInput: String?,
+        pathInput: String?,
+        labels: ArrayList<Int>?
+    ) {
         _createDialog.value = false
         val name = nameInput ?: java.lang.String.format(
             "%s_%s",
             res.getString(R.string.standard_name_recording),
             android.text.format.DateFormat.format(
-                "yyyy-MM-dd_hh-mm",
+                "yyyy-MM-dd_HH-mm",
                 Calendar.getInstance(Locale.getDefault())
             )
         )
         if (!validNameInput(name)) {
-            errorMessage = res.getString(R.string.dialog_invalid_name)
-            _createDialog.value = true
+            errorDialog(res.getString(R.string.dialog_invalid_name))
+            return
+        }
+        if (name.length > res.getInteger(R.integer.max_name_length)) {
+            errorDialog(res.getString(R.string.dialog_name_length))
             return
         }
 
@@ -202,18 +210,25 @@ class RecordViewModel(private val dataSource: Repository, application: Applicati
         }
         endRecordSession()
         File(tempFile).copyTo(newFile)
+
         val recordingDuration = getRecordingDuration() ?: currentRecordTime
         val audio =
-            EntryEntity(0, name, path, getDate(), recordingDuration)
-        saveRecordInDB(audio)
-
+            EntryEntity(
+                uid = 0,
+                recordingName = name,
+                recordingPath = path,
+                date = getDate(),
+                duration = recordingDuration
+            )
+        saveRecordInDB(audio, labels)
         File(tempFile).delete()
         resetView()
         errorMessage = null
     }
 
-    private fun saveRecordInDB(audio: EntryEntity) {
+    private fun saveRecordInDB(audio: EntryEntity, labels: ArrayList<Int>?) {
         val uid = dataSource.insert(audio).toInt()
+        if (labels != null) dataSource.insertRecLabels(LabelAssignmentEntity(0, uid, labels))
         if (markList.isNotEmpty()){
             saveMarksInDB(uid)
         }
@@ -221,7 +236,12 @@ class RecordViewModel(private val dataSource: Repository, application: Applicati
     }
 
     private fun validNameInput(name: String): Boolean {
-        return Pattern.compile("^[a-zA-Z0-9_-]+$").matcher(name).matches()
+        return Pattern.compile("^[a-zA-Z0-9_{}-]+$").matcher(name).matches()
+    }
+
+    private fun errorDialog(mes: String) {
+        errorMessage = mes
+        _createDialog.value = true
     }
 
     private fun getRecordingDuration(): String? {
